@@ -8,8 +8,8 @@ setup(name='pyHS100',
       author_email='sean@gadgetreactor.com',
       license='GPLv3',
       packages=['pyHS100'],
-      install_requires=['click', 'click-datetime', 'typing'],
-      python_requires='>=3.4',
+      install_requires=['click', 'typing', 'deprecation'],
+      python_requires='>=3.5',
       entry_points={
             'console_scripts': [
                   'pyhs100=pyHS100.cli:cli',
