@@ -1,36 +1,49 @@
+"""pyHS100 cli tool."""
 import sys
 import click
 import logging
-from click_datetime import Datetime
 from pprint import pformat as pf
 
 if sys.version_info < (3, 4):
-    print("To use this script you need python 3.4 or newer! got %s" %
-          sys.version_info)
+    print("To use this script you need python 3.4 or newer! got %s" % sys.version_info)
     sys.exit(1)
 
-from pyHS100 import (SmartDevice,
-                     SmartPlug,
-                     SmartBulb,
-                     SmartStrip,
-                     Discover)  # noqa: E402
+from pyHS100 import (
+    SmartDevice,
+    SmartPlug,
+    SmartBulb,
+    SmartStrip,
+    Discover,
+)  # noqa: E402
 
 pass_dev = click.make_pass_decorator(SmartDevice)
 
 
 @click.group(invoke_without_command=True)
-@click.option('--ip', envvar="PYHS100_IP", required=False,
-              help='The IP address of the device to connect to. This option '
-              'is deprecated and will be removed in the future; use --host '
-              'instead.')
-@click.option('--host', envvar="PYHS100_HOST", required=False,
-              help='The host name or IP address of the device to connect to.')
-@click.option('--alias', envvar="PYHS100_NAME", required=False,
-              help='The device name, or alias, of the device to connect to.')
-@click.option('--debug/--normal', default=False)
-@click.option('--bulb', default=False, is_flag=True)
-@click.option('--plug', default=False, is_flag=True)
-@click.option('--strip', default=False, is_flag=True)
+@click.option(
+    "--ip",
+    envvar="PYHS100_IP",
+    required=False,
+    help="The IP address of the device to connect to. This option "
+    "is deprecated and will be removed in the future; use --host "
+    "instead.",
+)
+@click.option(
+    "--host",
+    envvar="PYHS100_HOST",
+    required=False,
+    help="The host name or IP address of the device to connect to.",
+)
+@click.option(
+    "--alias",
+    envvar="PYHS100_NAME",
+    required=False,
+    help="The device name, or alias, of the device to connect to.",
+)
+@click.option("--debug/--normal", default=False)
+@click.option("--bulb", default=False, is_flag=True)
+@click.option("--plug", default=False, is_flag=True)
+@click.option("--strip", default=False, is_flag=True)
 @click.pass_context
 def cli(ctx, ip, host, alias, debug, bulb, plug, strip):
     """A cli tool for controlling TP-Link smart home plugs."""
@@ -46,8 +59,7 @@ def cli(ctx, ip, host, alias, debug, bulb, plug, strip):
         host = ip
 
     if alias is not None and host is None:
-        click.echo("Alias is given, using discovery to find host %s" %
-                   alias)
+        click.echo("Alias is given, using discovery to find host %s" % alias)
         host = find_host_from_alias(alias=alias)
         if host:
             click.echo("Found hostname is {}".format(host))
@@ -70,8 +82,7 @@ def cli(ctx, ip, host, alias, debug, bulb, plug, strip):
         elif strip:
             dev = SmartStrip(host)
         else:
-            click.echo(
-                "Unable to detect type, use --strip or --bulb or --plug!")
+            click.echo("Unable to detect type, use --strip or --bulb or --plug!")
             return
         ctx.obj = dev
 
@@ -80,15 +91,33 @@ def cli(ctx, ip, host, alias, debug, bulb, plug, strip):
 
 
 @cli.command()
-@click.option('--timeout', default=3, required=False)
-@click.option('--discover-only', default=False)
+@click.option("--save")
+def dump_discover(save):
+    for dev in Discover.discover(return_raw=True).values():
+        model = dev["system"]["get_sysinfo"]["model"]
+        hw_version = dev["system"]["get_sysinfo"]["hw_ver"]
+        save_to = "%s_%s.json" % (model, hw_version)
+        click.echo("Saving info to %s" % save_to)
+        with open(save_to, "w") as f:
+            import json
+
+            json.dump(dev, f, sort_keys=True, indent=4)
+
+
+@cli.command()
+@click.option("--timeout", default=3, required=False)
+@click.option("--discover-only", default=False)
+@click.option("--dump-raw", is_flag=True)
 @click.pass_context
-def discover(ctx, timeout, discover_only):
+def discover(ctx, timeout, discover_only, dump_raw):
     """Discover devices in the network."""
     click.echo("Discovering devices for %s seconds" % timeout)
-    found_devs = Discover.discover(timeout=timeout).items()
+    found_devs = Discover.discover(timeout=timeout, return_raw=dump_raw).items()
     if not discover_only:
         for ip, dev in found_devs:
+            if dump_raw:
+                click.echo(dev)
+                continue
             ctx.obj = dev
             ctx.invoke(state)
             print()
@@ -97,10 +126,12 @@ def discover(ctx, timeout, discover_only):
 
 
 def find_host_from_alias(alias, timeout=1, attempts=3):
-    """Discover a device identified by its alias"""
+    """Discover a device identified by its alias."""
     host = None
-    click.echo("Trying to discover %s using %s attempts of %s seconds" %
-               (alias, attempts, timeout))
+    click.echo(
+        "Trying to discover %s using %s attempts of %s seconds"
+        % (alias, attempts, timeout)
+    )
     for attempt in range(1, attempts):
         click.echo("Attempt %s of %s" % (attempt, attempts))
         found_devs = Discover.discover(timeout=timeout).items()
@@ -124,20 +155,26 @@ def sysinfo(dev):
 @click.pass_context
 def state(ctx, dev):
     """Print out device state and versions."""
-    click.echo(click.style("== %s - %s ==" % (dev.alias, dev.model),
-                           bold=True))
+    click.echo(click.style("== %s - %s ==" % (dev.alias, dev.model), bold=True))
 
-    click.echo(click.style("Device state: %s" % "ON" if dev.is_on else "OFF",
-                           fg="green" if dev.is_on else "red"))
+    click.echo(
+        click.style(
+            "Device state: %s" % "ON" if dev.is_on else "OFF",
+            fg="green" if dev.is_on else "red",
+        )
+    )
     if dev.num_children > 0:
         is_on = dev.is_on()
         aliases = dev.get_alias()
         for child in range(dev.num_children):
             click.echo(
-                click.style("  * %s state: %s" %
-                            (aliases[child],
-                             "ON" if is_on[child] else "OFF"),
-                            fg="green" if is_on[child] else "red"))
+                click.style(
+                    "  * %s state: %s"
+                    % (aliases[child], "ON" if is_on[child] else "OFF"),
+                    fg="green" if is_on[child] else "red",
+                )
+            )
+
     click.echo("Host/IP: %s" % dev.host)
     for k, v in dev.state_information.items():
         click.echo("%s: %s" % (k, v))
@@ -153,7 +190,7 @@ def state(ctx, dev):
 
 @cli.command()
 @pass_dev
-@click.argument('new_alias', required=False, default=None)
+@click.argument("new_alias", required=False, default=None)
 def alias(dev, new_alias):
     """Get or set the device alias."""
     if new_alias is not None:
@@ -165,11 +202,24 @@ def alias(dev, new_alias):
 
 @cli.command()
 @pass_dev
-@click.option('--year', type=Datetime(format='%Y'),
-              default=None, required=False)
-@click.option('--month', type=Datetime(format='%Y-%m'),
-              default=None, required=False)
-@click.option('--erase', is_flag=True)
+@click.argument("module")
+@click.argument("command")
+@click.argument("parameters", default=None, required=False)
+def raw_command(dev: SmartDevice, module, command, parameters):
+    """Run a raw command on the device."""
+    import ast
+
+    if parameters is not None:
+        parameters = ast.literal_eval(parameters)
+    res = dev._query_helper(module, command, parameters)
+    click.echo(res)
+
+
+@cli.command()
+@pass_dev
+@click.option("--year", type=click.DateTime(["%Y"]), default=None, required=False)
+@click.option("--month", type=click.DateTime(["%Y-%m"]), default=None, required=False)
+@click.option("--erase", is_flag=True)
 def emeter(dev, year, month, erase):
     """Query emeter for historical consumption."""
     click.echo(click.style("== Emeter ==", bold=True))
@@ -187,8 +237,7 @@ def emeter(dev, year, month, erase):
         emeter_status = dev.get_emeter_monthly(year.year)
     elif month:
         click.echo("== For month %s of %s ==" % (month.month, month.year))
-        emeter_status = dev.get_emeter_daily(year=month.year,
-                                             month=month.month)
+        emeter_status = dev.get_emeter_daily(year=month.year, month=month.month)
     else:
         emeter_status = dev.get_emeter_realtime()
         click.echo("== Current State ==")
@@ -201,9 +250,8 @@ def emeter(dev, year, month, erase):
 
 
 @cli.command()
-@click.argument("brightness", type=click.IntRange(0, 100), default=None,
-                required=False)
-@click.option('--transition', type=int, default=None, required=False,
+@click.argument("brightness", type=click.IntRange(0, 100), default=None, required=False)
+@click.option('--transition', type=int, default=-1, required=False,
               help='Transition period in milliseconds, default 500.')
 @pass_dev
 def brightness(dev, brightness, transition):
@@ -215,62 +263,51 @@ def brightness(dev, brightness, transition):
         click.echo("Brightness: %s" % dev.brightness)
     else:
         click.echo("Setting brightness to %s" % brightness)
-        if transition:
-            dev.set_brightness(brightness, transition)
-        else:
-            dev.brightness = brightness
-
+        dev.set_brightness(brightness, transition)
 
 @cli.command()
-@click.argument("temperature", type=click.IntRange(2500, 9000), default=None,
-                required=False)
-@click.option('--transition', type=int, default=None, required=False,
+@click.argument("temperature", type=click.IntRange(2500, 9000), default=None, required=False)
+@click.option('--transition', type=int, default=-1, required=False,
               help='Transition period in milliseconds, default 500.')
 @pass_dev
-def temperature(dev, temperature, transition):
+def temperature(dev: SmartBulb, temperature, transition):
     """Get or set color temperature. (Bulb only)"""
     if temperature is None:
         click.echo("Color temperature: %s" % dev.color_temp)
         if dev.valid_temperature_range != (0, 0):
             click.echo("(min: %s, max: %s)" % dev.valid_temperature_range)
         else:
-            click.echo("Temperature range unknown, please open a github issue"
-                       " or a pull request for model '%s'" % dev.model)
+            click.echo(
+                "Temperature range unknown, please open a github issue"
+                " or a pull request for model '%s'" % dev.model
+            )
     else:
         click.echo("Setting color temperature to %s" % temperature)
-        if transition:
-            dev.set_color_temp(temperature, transition)
-        else:
-            dev.color_temp = temperature
-
+        dev.set_color_temp(temperature, transition)
 
 @cli.command()
 @click.argument("h", type=click.IntRange(0, 360), default=None, required=False)
 @click.argument("s", type=click.IntRange(0, 100), default=None, required=False)
 @click.argument("v", type=click.IntRange(0, 100), default=None, required=False)
-@click.option('--transition', type=int, default=None, required=False,
+@click.option('--transition', type=int, default=-1, required=False,
               help='Transition period in milliseconds, default 500.')
 @click.pass_context
 @pass_dev
 def hsv(dev, ctx, h, s, v, transition):
     """Get or set color in HSV. (Bulb only)"""
-    if h is None:
+    if h is None or s is None or v is None:
         click.echo("Current HSV: %s %s %s" % dev.hsv)
     elif s is None or v is None:
         raise click.BadArgumentUsage("Setting a color requires 3 values.", ctx)
     else:
         click.echo("Setting HSV: %s %s %s" % (h, s, v))
-        if transition:
-            dev.set_hsv((h, s, v), transition)
-        else:
-            dev.hsv = (h, s, v)
-
+        dev.set_hsv((h, s, v), transition)
 
 @cli.command()
-@click.argument('state', type=bool, required=False)
+@click.argument("state", type=bool, required=False)
 @pass_dev
 def led(dev, state):
-    """Get or set led state. (Plug only)"""
+    """Get or set (Plug's) led state."""
     if state is not None:
         click.echo("Turning led to %s" % state)
         dev.led = state
@@ -286,7 +323,7 @@ def time(dev):
 
 
 @cli.command()
-@click.argument('index', type=int, required=False)
+@click.argument("index", type=int, required=False)
 @pass_dev
 def on(plug, index):
     """Turn the device on."""
@@ -298,7 +335,7 @@ def on(plug, index):
 
 
 @cli.command()
-@click.argument('index', type=int, required=False)
+@click.argument("index", type=int, required=False)
 @pass_dev
 def off(plug, index):
     """Turn the device off."""
