@@ -1,14 +1,20 @@
-from pyHS100 import SmartDevice, SmartDeviceException
+from pyHS100 import DeviceType, SmartDevice, SmartDeviceException
+from .protocol import TPLinkSmartHomeProtocol
+from deprecation import deprecated
 import re
-from typing import Any, Dict, Optional, Tuple
+from datetime import datetime
+from typing import Any, Dict, Tuple
 
-TPLINK_KELVIN = {'LB130': (2500, 9000),
-                 'LB120': (2700, 6500),
-                 'LB230': (2500, 9000),
-                 'KB130': (2500, 9000),
-                 'KL130': (2500, 9000),
-                 'KL120\(EU\)': (2700, 6500),
-                 'KL120\(US\)': (2700, 5000)}
+
+TPLINK_KELVIN = {
+    "LB130": (2500, 9000),
+    "LB120": (2700, 6500),
+    "LB230": (2500, 9000),
+    "KB130": (2500, 9000),
+    "KL130": (2500, 9000),
+    "KL120\(EU\)": (2700, 6500),
+    "KL120\(US\)": (2700, 5000),
+}
 
 
 class SmartBulb(SmartDevice):
@@ -16,88 +22,90 @@ class SmartBulb(SmartDevice):
 
     Usage example when used as library:
     p = SmartBulb("192.168.1.105")
+
     # print the devices alias
     print(p.alias)
+
     # change state of bulb
-    p.state = "ON"
-    p.state = "OFF"
-    # set transition period
-    p.transition_period = 1000
+    p.turn_on()
+    p.turn_off()
+
     # query and print current state of plug
     print(p.state)
+
     # check whether the bulb supports color changes
     if p.is_color:
+
     # set the color to an HSV tuple
-    p.hsv = (180, 100, 100)
+    p.set_hsv(180, 100, 100)
     # get the current HSV value
     print(p.hsv)
+
     # check whether the bulb supports setting color temperature
     if p.is_variable_color_temp:
     # set the color temperature in Kelvin
-    p.color_temp = 3000
+    p.set_color_temp(3000)
     # get the current color temperature
     print(p.color_temp)
+
     # check whether the bulb is dimmable
     if p.is_dimmable:
     # set the bulb to 50% brightness
-    p.brightness = 50
+    p.set_brightness(50)
     # check the current brightness
     print(p.brightness)
 
     Errors reported by the device are raised as SmartDeviceExceptions,
     and should be handled by the user of the library.
-
     """
-    # bulb states
-    BULB_STATE_ON = 'ON'
-    BULB_STATE_OFF = 'OFF'
 
-    # default transition period
-    DEFAULT_TRANSITION_PERIOD = 500
+    LIGHT_SERVICE = "smartlife.iot.smartbulb.lightingservice"
 
-    def __init__(self,
-                 host: str,
-                 protocol: 'TPLinkSmartHomeProtocol' = None) -> None:
-        SmartDevice.__init__(self, host, protocol)
+    def __init__(
+        self,
+        host: str,
+        protocol: TPLinkSmartHomeProtocol = None,
+        context: str = None,
+        cache_ttl: int = 3,
+    ) -> None:
+        SmartDevice.__init__(
+            self, host=host, protocol=protocol, context=context, cache_ttl=cache_ttl
+        )
         self.emeter_type = "smartlife.iot.common.emeter"
-        self.transition_period = self.DEFAULT_TRANSITION_PERIOD
+        self._device_type = DeviceType.Bulb
+        self._transition_period = 500
 
     @property
     def is_color(self) -> bool:
-        """
-        Whether the bulb supports color changes
+        """Whether the bulb supports color changes.
 
         :return: True if the bulb supports color changes, False otherwise
         :rtype: bool
         """
-        return bool(self.sys_info['is_color'])
+        return bool(self.sys_info["is_color"])
 
     @property
     def is_dimmable(self) -> bool:
-        """
-        Whether the bulb supports brightness changes
+        """Whether the bulb supports brightness changes.
 
         :return: True if the bulb supports brightness changes, False otherwise
         :rtype: bool
         """
-        return bool(self.sys_info['is_dimmable'])
+        return bool(self.sys_info["is_dimmable"])
 
     @property
     def is_variable_color_temp(self) -> bool:
-        """
-        Whether the bulb supports color temperature changes
+        """Whether the bulb supports color temperature changes.
 
         :return: True if the bulb supports color temperature changes, False
         otherwise
         :rtype: bool
         """
-        return bool(self.sys_info['is_variable_color_temp'])
+        return bool(self.sys_info["is_variable_color_temp"])
 
     @property
     def valid_temperature_range(self) -> Tuple[int, int]:
-        """
-        Returns the white temperature range (in Kelvin)
-        depending on the bulb model
+        """Return the device-specific white temperature range (in Kelvin).
 
         :return: White temperature range in Kelvin (minimun, maximum)
         :rtype: tuple
@@ -105,142 +113,150 @@ class SmartBulb(SmartDevice):
         if not self.is_variable_color_temp:
             return (0, 0)
         for model, temp_range in TPLINK_KELVIN.items():
-            if re.match(model, self.sys_info['model']):
+            if re.match(model, self.sys_info["model"]):
                 return temp_range
         return (0, 0)
 
+    @property
+    def transition_period(self) -> int:
+        """Return default transition period.
+
+        :return: transition period in milliseconds
+        :rtype: int
+        """
+        return self._transition_period
+
+    @transition_period.setter
+    def transition_period(self, period: int):
+        """Set default transition period.
+
+        :param int period: transition period in milliseconds, default 500
+        """
+
+        if period < 0:
+            raise SmartDeviceException(
+                'Invalid transition value: {}'
+                '(requires positive int value)'.format(period))
+
+        self._transition_period = period
+
     def get_light_state(self) -> Dict:
-        return self._query_helper("smartlife.iot.smartbulb.lightingservice",
-                                  "get_light_state")
+        """Query the light state."""
+        return self._query_helper(self.LIGHT_SERVICE, "get_light_state")
 
     def set_light_state(self, state: Dict) -> Dict:
-        return self._query_helper("smartlife.iot.smartbulb.lightingservice",
-                                  "transition_light_state", state)
+        """Set the light state."""
+        return self._query_helper(self.LIGHT_SERVICE, "transition_light_state", state)
 
     @property
-    def hsv(self) -> Optional[Tuple[int, int, int]]:
-        """
-        Returns the current HSV state of the bulb, if supported
+    def hsv(self) -> Tuple[int, int, int]:
+        """Return the current HSV state of the bulb.
 
         :return: hue, saturation and value (degrees, %, %)
         :rtype: tuple
         """
 
         if not self.is_color:
-            return None
+            raise SmartDeviceException("Bulb does not support color.")
 
         light_state = self.get_light_state()
         if not self.is_on:
-            hue = light_state['dft_on_state']['hue']
-            saturation = light_state['dft_on_state']['saturation']
-            value = light_state['dft_on_state']['brightness']
+            hue = light_state["dft_on_state"]["hue"]
+            saturation = light_state["dft_on_state"]["saturation"]
+            value = light_state["dft_on_state"]["brightness"]
         else:
-            hue = light_state['hue']
-            saturation = light_state['saturation']
-            value = light_state['brightness']
+            hue = light_state["hue"]
+            saturation = light_state["saturation"]
+            value = light_state["brightness"]
 
         return hue, saturation, value
 
-    @hsv.setter
+    @hsv.setter  # type: ignore
+    @deprecated(details="Use set_hsv()")
     def hsv(self, state: Tuple[int, int, int]):
-        """
-        Sets new HSV, if supported
+        return self.set_hsv(state[0], state[1], state[2])
+
+    def _raise_for_invalid_brightness(self, value):
+        if not isinstance(value, int) or not (0 <= value <= 100):
+            raise ValueError(
+                "Invalid brightness value: {} " "(valid range: 0-100%)".format(value)
+            )
+
+    def set_hsv(self, hue: int, saturation: int, value: int, period: int = -1):
+        """Set new HSV.
 
         :param tuple state: hue, saturation and value (degrees, %, %)
+        :param int period: transition period in milliseconds
         """
-        self.set_hsv(state)
-
-    def set_hsv(self, state: Tuple[int, int, int], period=None):
-        """
-        Sets new HSV, if supported
-
-        :param tuple state: hue, saturation and value (degrees, %, %)
-        :param period: The transition period in milliseconds, default 500
-        """
-
         if not self.is_color:
-            return None
+            raise SmartDeviceException("Bulb does not support color.")
 
-        if not isinstance(state[0], int) or not (0 <= state[0] <= 360):
-            raise SmartDeviceException(
-                    'Invalid hue value: {} '
-                    '(valid range: 0-360)'.format(state[0]))
+        if not isinstance(hue, int) or not (0 <= hue <= 360):
+            raise ValueError(
+                "Invalid hue value: {} " "(valid range: 0-360)".format(hue)
+            )
 
-        if not isinstance(state[1], int) or not (0 <= state[1] <= 100):
-            raise SmartDeviceException(
-                    'Invalid saturation value: {} '
-                    '(valid range: 0-100%)'.format(state[1]))
-
-        if not isinstance(state[2], int) or not (0 <= state[2] <= 100):
-            raise SmartDeviceException(
-                    'Invalid brightness value: {} '
-                    '(valid range: 0-100%)'.format(state[2]))
-
-        if period is None:
-            period = self.transition_period
+        if not isinstance(saturation, int) or not (0 <= saturation <= 100):
+            raise ValueError(
+                "Invalid saturation value: {} "
+                "(valid range: 0-100%)".format(saturation)
+            )
 
         if period < 0:
-            raise SmartDeviceException(
-                'Invalid transition value: {}'
-                '(requires positive value)'.format(period))
+            period = self._transition_period
+
+        self._raise_for_invalid_brightness(value)
 
         light_state = {
-            "hue": state[0],
-            "saturation": state[1],
-            "brightness": state[2],
+            "hue": hue,
+            "saturation": saturation,
+            "brightness": value,
             "color_temp": 0,
             "transition_period": period
-            }
+        }
         self.set_light_state(light_state)
 
     @property
-    def color_temp(self) -> Optional[int]:
-        """
-        Color temperature of the device, if supported
+    def color_temp(self) -> int:
+        """Return color temperature of the device.
 
         :return: Color temperature in Kelvin
         :rtype: int
         """
         if not self.is_variable_color_temp:
-            return None
+            raise SmartDeviceException("Bulb does not support colortemp.")
 
         light_state = self.get_light_state()
         if not self.is_on:
-            return int(light_state['dft_on_state']['color_temp'])
+            return int(light_state["dft_on_state"]["color_temp"])
         else:
-            return int(light_state['color_temp'])
+            return int(light_state["color_temp"])
 
-    @color_temp.setter
+    @color_temp.setter  # type: ignore
+    @deprecated(details="use set_color_temp")
     def color_temp(self, temp: int) -> None:
-        """
-        Set the color temperature of the device, if supported
-
-        :param int temp: The new color temperature, in Kelvin
-        """
         self.set_color_temp(temp)
 
-    def set_color_temp(self, temp: int, period=None) -> None:
-        """
-        Set the color temperature of the device, if supported
+    def set_color_temp(self, temp: int, period: int = -1) -> None:
+        """Set the color temperature of the device.
 
         :param int temp: The new color temperature, in Kelvin
-        :param int period: The transition period in milliseconds, default 500
+        :param int period: transition period in milliseconds
         """
         if not self.is_variable_color_temp:
-            return None
+            raise SmartDeviceException("Bulb does not support colortemp.")
 
-        if temp < self.valid_temperature_range[0] or \
-                temp > self.valid_temperature_range[1]:
-            raise ValueError("Temperature should be between {} "
-                             "and {}".format(*self.valid_temperature_range))
-
-        if period is None:
-            period = self.transition_period
+        if (
+            temp < self.valid_temperature_range[0]
+            or temp > self.valid_temperature_range[1]
+        ):
+            raise ValueError(
+                "Temperature should be between {} "
+                "and {}".format(*self.valid_temperature_range)
+            )
 
         if period < 0:
-            raise SmartDeviceException(
-                'Invalid transition value: {}'
-                '(requires positive value)'.format(period))
+            period = self._transition_period
 
         light_state = {
             "color_temp": temp,
@@ -249,49 +265,39 @@ class SmartBulb(SmartDevice):
         self.set_light_state(light_state)
 
     @property
-    def brightness(self) -> Optional[int]:
-        """
-        Current brightness of the device, if supported
+    def brightness(self) -> int:
+        """Current brightness of the device.
 
         :return: brightness in percent
         :rtype: int
         """
-        if not self.is_dimmable:
-            return None
+        if not self.is_dimmable:  # pragma: no cover
+            raise SmartDeviceException("Bulb is not dimmable.")
 
         light_state = self.get_light_state()
         if not self.is_on:
-            return int(light_state['dft_on_state']['brightness'])
+            return int(light_state["dft_on_state"]["brightness"])
         else:
-            return int(light_state['brightness'])
+            return int(light_state["brightness"])
 
-    @brightness.setter
+    @brightness.setter  # type: ignore
+    @deprecated(details="use set_brightness")
     def brightness(self, brightness: int) -> None:
-        """
-        Set the current brightness of the device, if supported
-
-        :param int brightness: brightness in percent
-        """
         self.set_brightness(brightness)
 
-    def set_brightness(self, brightness: int, period=None) -> None:
-        """
-        Set the current brightness of the device, if supported
+    def set_brightness(self, brightness: int, period: int = -1) -> None:
+        """Set the current brightness of the device.
 
         :param int brightness: brightness in percent
-        :param int period: The transition period in milliseconds, default 500
+        :param int period: transition period in milliseconds
         """
+        if not self.is_dimmable:  # pragma: no cover
+            raise SmartDeviceException("Bulb is not dimmable.")
 
-        if not self.is_dimmable:
-            return None
-
-        if period is None:
-            period = self.transition_period
+        self._raise_for_invalid_brightness(brightness)
 
         if period < 0:
-            raise SmartDeviceException(
-                'Invalid transition value: {}'
-                '(requires positive value)'.format(period))
+            period = self._transition_period
 
         light_state = {
             "brightness": brightness,
@@ -299,72 +305,50 @@ class SmartBulb(SmartDevice):
         }
         self.set_light_state(light_state)
 
-    @property
+    @property  # type: ignore
+    @deprecated(details="use is_on() and is_off()")
     def state(self) -> str:
-        """
-        Retrieve the bulb state
+        """Retrieve the bulb state.
 
         :returns: one of
-                  BULB_STATE_ON
-                  BULB_STATE_OFF
+                  STATE_ON
+                  STATE_OFF
         :rtype: str
         """
-        light_state = self.get_light_state()
-        if light_state['on_off']:
-            return self.BULB_STATE_ON
-        return self.BULB_STATE_OFF
+        if self.is_on:
+            return self.STATE_ON
 
-    @state.setter
+        return self.STATE_OFF
+
+    @state.setter  # type: ignore
+    @deprecated(details="use turn_on() and turn_off()")
     def state(self, bulb_state: str) -> None:
-        """
-        Set the new bulb state
+        """Set the new bulb state.
 
         :param bulb_state: one of
-                           BULB_STATE_ON
-                           BULB_STATE_OFF
+                           STATE_ON
+                           STATE_OFF
         """
-        self.set_state(bulb_state)
-
-    def set_state(self, bulb_state: str, period=None) -> None:
-        """
-        Set the new bulb state
-
-        :param bulb_state: one of
-                           BULB_STATE_ON
-                           BULB_STATE_OFF
-        :param int period: The transition period in milliseconds, default 500
-        """
-        if bulb_state == self.BULB_STATE_ON:
+        if bulb_state == self.STATE_ON:
             new_state = 1
-        elif bulb_state == self.BULB_STATE_OFF:
+        elif bulb_state == self.STATE_OFF:
             new_state = 0
         else:
             raise ValueError
 
-        if period is None:
-            period = self.transition_period
-
-        if period < 0:
-            raise SmartDeviceException(
-                'Invalid transition value: {}'
-                '(requires positive value)'.format(period))
-
-        light_state = {
-            "on_off": new_state,
-            "transition_period": period
-        }
+        light_state = {"on_off": new_state}
         self.set_light_state(light_state)
 
     @property
     def state_information(self) -> Dict[str, Any]:
-        """
-        Return bulb-specific state information.
+        """Return bulb-specific state information.
+
         :return: Bulb information dict, keys in user-presentable form.
         :rtype: dict
         """
         info = {
-            'Brightness': self.brightness,
-            'Is dimmable': self.is_dimmable,
+            "Brightness": self.brightness,
+            "Is dimmable": self.is_dimmable,
         }  # type: Dict[str, Any]
         if self.is_variable_color_temp:
             info["Color temperature"] = self.color_temp
@@ -376,19 +360,39 @@ class SmartBulb(SmartDevice):
 
     @property
     def is_on(self) -> bool:
-        return bool(self.state == self.BULB_STATE_ON)
+        """Return whether the device is on."""
+        light_state = self.get_light_state()
+        return bool(light_state["on_off"])
 
-    def turn_off(self, period=None) -> None:
-        """
-        Turn the bulb off.
-        """
-        self.set_state(self.BULB_STATE_OFF, period)
+    def turn_off(self, period: int = -1) -> None:
+        """Turn the bulb off.
 
-    def turn_on(self, period=None) -> None:
+        :param int period: transition period in milliseconds
         """
-        Turn the bulb on.
+
+        if period < 0:
+            period = self._transition_period
+
+        light_state = {
+            "on_off": 0,
+            "transition_period": period
+        }
+        self.set_light_state(light_state)
+
+    def turn_on(self, period: int = -1) -> None:
+        """Turn the bulb on.
+
+        :param int period: transition period in milliseconds
         """
-        self.set_state(self.BULB_STATE_ON, period)
+
+        if period < 0:
+            period = self._transition_period
+
+        light_state = {
+            "on_off": 1,
+            "transition_period": period
+        }
+        self.set_light_state(light_state)
 
     @property
     def has_emeter(self) -> bool:
